\name{Ssa.RefSeq.db-package}
\alias{Ssa.RefSeq.db-package}
\alias{Ssa.RefSeq.db}
\docType{package}
\title{
	The package provides gene information for the A.salmon genome
}
\description{
  The package provides mappings between A.salmon gene IDs and
  annotations (gene name, GO) and gene locations. All mappings and
  annotations are based on the latest A.salmon genome assambly
  (ICSASG_v2).
  The package provides all annotation data (as .sqlite DB) as well as
  functions to query the DB.
}
\details{
\tabular{ll}{
Package: \tab Ssa.RefSeq.db\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-02-17\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Fabian Grammes
Maintainer: Fabian Grammes <fabian.grammes@nmbu.no>
}

